# openapi.yaml
openapi: 3.0.0
info:
  title: SPK
  version: 1.0.0
  description: ini untuk spk
servers:
  - url: '{{BASE_URL}}'
  - url: '{{PROCESSOR_URL}}'
paths:
...
  /criteria/bulk:
    post:
      summary: criteria bulk
      responses:
        '200':
          description: '200'
          headers:
            Content-Security-Policy:
              schema:
                type: string
              example: >-
                default-src 'self';base-uri 'self';font-src 'self' https:
                data:;form-action 'self';frame-ancestors 'self';img-src 'self'
                data:;object-src 'none';script-src 'self';script-src-attr
                'none';style-src 'self' https:
                'unsafe-inline';upgrade-insecure-requests
            Cross-Origin-Opener-Policy:
              schema:
                type: string
              example: same-origin
            Origin-Agent-Cluster:
              schema:
                type: string
              example: '?1'
            Referrer-Policy:
              schema:
                type: string
              example: no-referrer
            Strict-Transport-Security:
              schema:
                type: string
              example: max-age=31536000; includeSubDomains
            X-Content-Type-Options:
              schema:
                type: string
              example: nosniff
            X-DNS-Prefetch-Control:
              schema:
                type: string
              example: 'off'
            X-Download-Options:
              schema:
                type: string
              example: noopen
            X-Frame-Options:
              schema:
                type: string
              example: SAMEORIGIN
            X-Permitted-Cross-Domain-Policies:
              schema:
                type: string
              example: none
            X-XSS-Protection:
              schema:
                type: integer
              example: 0
            Vary:
              schema:
                type: string
              example: Origin
            Access-Control-Allow-Credentials:
              schema:
                type: boolean
              example: true
            Content-Length:
              schema:
                type: integer
              example: 44
            ETag:
              schema:
                type: string
              example: W/"2c-ml+GcHRkrdOT4M9jrfVedL4t1Tc"
            Date:
              schema:
                type: string
              example: Thu, 14 May 2026 19:45:34 GMT
            Connection:
              schema:
                type: string
              example: keep-alive
            Keep-Alive:
              schema:
                type: string
              example: timeout=5
        '400':
          description: '400'
          headers:
            Content-Security-Policy:
              schema:
                type: string
              example: >-
                default-src 'self';base-uri 'self';font-src 'self' https:
                data:;form-action 'self';frame-ancestors 'self';img-src 'self'
                data:;object-src 'none';script-src 'self';script-src-attr
                'none';style-src 'self' https:
                'unsafe-inline';upgrade-insecure-requests
            Cross-Origin-Opener-Policy:
              schema:
                type: string
              example: same-origin
            Origin-Agent-Cluster:
              schema:
                type: string
              example: '?1'
            Referrer-Policy:
              schema:
                type: string
              example: no-referrer
            Strict-Transport-Security:
              schema:
                type: string
              example: max-age=31536000; includeSubDomains
            X-Content-Type-Options:
              schema:
                type: string
              example: nosniff
            X-DNS-Prefetch-Control:
              schema:
                type: string
              example: 'off'
            X-Download-Options:
              schema:
                type: string
              example: noopen
            X-Frame-Options:
              schema:
                type: string
              example: SAMEORIGIN
            X-Permitted-Cross-Domain-Policies:
              schema:
                type: string
              example: none
            X-XSS-Protection:
              schema:
                type: integer
              example: 0
            Vary:
              schema:
                type: string
              example: Origin
            Access-Control-Allow-Credentials:
              schema:
                type: boolean
              example: true
            Content-Length:
              schema:
                type: integer
              example: 71
            ETag:
              schema:
                type: string
              example: W/"47-AxbBiD9nw2yUPfnQY+eLIaH5Uzo"
            Date:
              schema:
                type: string
              example: Thu, 14 May 2026 19:22:43 GMT
            Connection:
              schema:
                type: string
              example: keep-alive
            Keep-Alive:
              schema:
                type: string
              example: timeout=5
      tags:
        - 'TOPSIS data '
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  weight:
                    type: number
                  type:
                    type: string
            example:
              - name: ipk
                weight: 0.3
                type: benefit
              - name: penghasilan_ortu
                weight: 0.2
                type: cost
              - name: jumlah_tanggungan
                weight: 0.1
                type: cost
              - name: skor_prestasi
                weight: 0.2
                type: benefit
              - name: keaktifan_organisasi
                weight: 0.2
                type: benefit
    patch:
      summary: patch criteria bulk
      responses:
        '404':
          description: '404'
          headers:
            Content-Security-Policy:
              schema:
                type: string
              example: >-
                default-src 'self';base-uri 'self';font-src 'self' https:
                data:;form-action 'self';frame-ancestors 'self';img-src 'self'
                data:;object-src 'none';script-src 'self';script-src-attr
                'none';style-src 'self' https:
                'unsafe-inline';upgrade-insecure-requests
            Cross-Origin-Opener-Policy:
              schema:
                type: string
              example: same-origin
            Origin-Agent-Cluster:
              schema:
                type: string
              example: '?1'
            Referrer-Policy:
              schema:
                type: string
              example: no-referrer
            Strict-Transport-Security:
              schema:
                type: string
              example: max-age=31536000; includeSubDomains
            X-Content-Type-Options:
              schema:
                type: string
              example: nosniff
            X-DNS-Prefetch-Control:
              schema:
                type: string
              example: 'off'
            X-Download-Options:
              schema:
                type: string
              example: noopen
            X-Frame-Options:
              schema:
                type: string
              example: SAMEORIGIN
            X-Permitted-Cross-Domain-Policies:
              schema:
                type: string
              example: none
            X-XSS-Protection:
              schema:
                type: integer
              example: 0
            Vary:
              schema:
                type: string
              example: Origin
            Access-Control-Allow-Credentials:
              schema:
                type: boolean
              example: true
            Content-Length:
              schema:
                type: integer
              example: 74
            ETag:
              schema:
                type: string
              example: W/"4a-g1iFWXPrLTPqRO11sUY/PuYKZDU"
            Date:
              schema:
                type: string
              example: Thu, 14 May 2026 19:47:46 GMT
            Connection:
              schema:
                type: string
              example: keep-alive
            Keep-Alive:
              schema:
                type: string
              example: timeout=5
      tags:
        - 'TOPSIS data '
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  weight:
                    type: number
                  type:
                    type: string
            example:
              - name: ipk
                weight: 0.3
                type: BENEFIT
              - name: penghasilan_ortu
                weight: 0.2
                type: COST
              - name: jumlah_tanggungan
                weight: 0.1
                type: COST
              - name: skor_prestasi
                weight: 0.1
                type: BENEFIT
              - name: keaktifan_organisasi
                weight: 0.1
                type: BENEFIT
              - name: semester
                weight: 0.2
                type: COST
    delete:
      summary: delete criteria bulk
      responses:
        '200':
          description: '200'
          headers:
            Content-Security-Policy:
              schema:
                type: string
              example: >-
                default-src 'self';base-uri 'self';font-src 'self' https:
                data:;form-action 'self';frame-ancestors 'self';img-src 'self'
                data:;object-src 'none';script-src 'self';script-src-attr
                'none';style-src 'self' https:
                'unsafe-inline';upgrade-insecure-requests
            Cross-Origin-Opener-Policy:
              schema:
                type: string
              example: same-origin
            Origin-Agent-Cluster:
              schema:
                type: string
              example: '?1'
            Referrer-Policy:
              schema:
                type: string
              example: no-referrer
            Strict-Transport-Security:
              schema:
                type: string
              example: max-age=31536000; includeSubDomains
            X-Content-Type-Options:
              schema:
                type: string
              example: nosniff
            X-DNS-Prefetch-Control:
              schema:
                type: string
              example: 'off'
            X-Download-Options:
              schema:
                type: string
              example: noopen
            X-Frame-Options:
              schema:
                type: string
              example: SAMEORIGIN
            X-Permitted-Cross-Domain-Policies:
              schema:
                type: string
              example: none
            X-XSS-Protection:
              schema:
                type: integer
              example: 0
            Vary:
              schema:
                type: string
              example: Origin
            Access-Control-Allow-Credentials:
              schema:
                type: boolean
              example: true
            Content-Length:
              schema:
                type: integer
              example: 42
            ETag:
              schema:
                type: string
              example: W/"2a-MwGbTgZBKFtANb91E+VNlyojU+4"
            Date:
              schema:
                type: string
              example: Thu, 14 May 2026 19:46:01 GMT
            Connection:
              schema:
                type: string
              example: keep-alive
            Keep-Alive:
              schema:
                type: string
              example: timeout=5
        '404':
          description: '404'
          headers:
            Content-Security-Policy:
              schema:
                type: string
              example: >-
                default-src 'self';base-uri 'self';font-src 'self' https:
                data:;form-action 'self';frame-ancestors 'self';img-src 'self'
                data:;object-src 'none';script-src 'self';script-src-attr
                'none';style-src 'self' https:
                'unsafe-inline';upgrade-insecure-requests
            Cross-Origin-Opener-Policy:
              schema:
                type: string
              example: same-origin
            Origin-Agent-Cluster:
              schema:
                type: string
              example: '?1'
            Referrer-Policy:
              schema:
                type: string
              example: no-referrer
            Strict-Transport-Security:
              schema:
                type: string
              example: max-age=31536000; includeSubDomains
            X-Content-Type-Options:
              schema:
                type: string
              example: nosniff
            X-DNS-Prefetch-Control:
              schema:
                type: string
              example: 'off'
            X-Download-Options:
              schema:
                type: string
              example: noopen
            X-Frame-Options:
              schema:
                type: string
              example: SAMEORIGIN
            X-Permitted-Cross-Domain-Policies:
              schema:
                type: string
              example: none
            X-XSS-Protection:
              schema:
                type: integer
              example: 0
            Vary:
              schema:
                type: string
              example: Origin
            Access-Control-Allow-Credentials:
              schema:
                type: boolean
              example: true
            Content-Length:
              schema:
                type: integer
              example: 144
            ETag:
              schema:
                type: string
              example: W/"90-o0PX/hmV0qQCzkaYFFScDH7wHBg"
            Date:
              schema:
                type: string
              example: Thu, 14 May 2026 19:52:00 GMT
            Connection:
              schema:
                type: string
              example: keep-alive
            Keep-Alive:
              schema:
                type: string
              example: timeout=5
      tags:
        - 'TOPSIS data '
      description: >-
        ## Delete Criteria Bulk


        Deletes multiple criteria at once by their names. This is useful for
        removing several criteria in a single request instead of deleting them
        one by one.


        ---


        ### Method & URL


        ```

        DELETE {{BASE_URL}}/criteria/bulk

        ```


        ---


        ### Request Body


        **Content-Type:** `application/json`


        | Field   | Type            | Required |
        Description                                      |

        |---------|-----------------|----------|--------------------------------------------------|

        | `names` | `array<string>` | ✅ Yes   | A list of criteria names to be
        deleted.          |


        **Example:**


        ```json

        {
            "names": [
                "ipk",
                "semester",
                "penghasilan_ortu",
                "jumlah_tanggungan",
                "keaktifan_organisasi",
                "skor_prestasi"
            ]
        }

        ```


        ---


        ### Responses


        #### ✅ 200 — Success


        All specified criteria were found and successfully deleted.


        ---


        #### ❌ 404 — Not Found


        None of the specified criteria were deleted. This typically means none
        of the provided names exist in the database.


        ```json

        {
            "message": "No criteria were deleted. Please check the names and try again. (maybe they don't exist)",
            "error": {
                "name": "HttpError",
                "status": 404
            }
        }

        ```


        ---


        ### Notes


        - The `names` array must contain at least one entry.

        - If **none** of the provided names match existing criteria, a `404`
        response is returned.

        - Criteria names are case-sensitive — make sure the names match exactly
        as stored in the database.

        - This endpoint performs a bulk operation; partial deletions (where only
        some names match) may or may not be supported depending on the server
        implementation — verify behavior with the backend if needed.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                names:
                  type: array
                  items:
                    type: string
            example:
              names:
                - ipk
                - semester
                - penghasilan_ortu
                - jumlah_tanggungan
                - keaktifan_organisasi
                - skor_prestasi
...